+++ saleor/graphql/payment/tests/mutations/test_transaction_event_report.py
@@ -935,6 +935,10 @@ def test_transaction_event_report_incorrect_amount_for_already_existing(
     error = transaction_report_data["errors"][0]
     assert error["code"] == TransactionEventReportErrorCode.INCORRECT_DETAILS.name
     assert error["field"] == "pspReference"
+    assert psp_reference in error["message"]
+    assert TransactionEventType.CHARGE_SUCCESS in error["message"]
+    assert str(already_existing_amount) in error["message"]
+    assert str(new_amount) in error["message"]
 
     assert TransactionEvent.objects.count() == 2
     event = TransactionEvent.objects.filter(
@@ -942,6 +946,7 @@ def test_transaction_event_report_incorrect_amount_for_already_existing(
     ).first()
     assert event
     assert event.include_in_calculations is False
+    assert event.message == error["message"]
 
 
 @patch(